Wordle: complete checkGuess() logic, show AlertDialog after game
Letters are now validated correctly, taking into account how often the respective letters occur in the targetWord, i.e. the number of green and yellow tiles of a given letter now equals the total number of occurrences of that letter in the targetWord.

After a won or lost game, an AlertDialog is shown, allowing the user to restart or end the game (return to GamesListActivity).